backend/requirements.txt
Django==5.0.6
django-cors-headers==4.3.1
django-environ==0.11.2
djangorestframework==3.15.1
flake8==7.0.0
psycopg==3.1.19
psycopg-binary==3.1.19
pytest==8.2.2
pytest-cov==5.0.0
pytest-django==4.8.0
pytest-mock==3.14.0